+++ R/compile_qmd_course.R
@@ -18,7 +18,8 @@
 #' @importFrom htmltools htmlTemplate renderDocument save_html
 #' @importFrom furrr future_map_lgl furrr_options
 #' @importFrom future plan
-#' @importFrom cli cli_alert_info cli_alert_warning
+#' @importFrom cli cli_alert_info cli_alert_warning cli_alert_success
+#' @importFrom progressr handlers progressor with_progress
 #'
 #' @return character. The path to the resulting html file
 #'
@@ -36,8 +37,19 @@
 #'   package = "squash"
 #' )
 #'
+#' # copy course tree in tmpdir, add quarto porject file
+#' tmp_course_path <- tempfile(pattern = "course")
+#' dir.create(tmp_course_path)
+#' file.create(file.path(tmp_course_path, "_quarto.yaml"))
+#'
+#' file.copy(
+#'   from = courses_path,
+#'   to = tmp_course_path,
+#'   recursive = TRUE
+#' )
+#'
 #' qmds <- list.files(
-#'   path = courses_path,
+#'   path = tmp_course_path,
 #'   full.names = TRUE,
 #'   recursive = TRUE,
 #'   pattern = "qmd$"
@@ -57,6 +69,7 @@
 #'
 #' # clean up
 #' unlink(temp_dir, recursive = TRUE)
+#' unlink(tmp_course_path, recursive = TRUE)
 compile_qmd_course <- function(
     vec_qmd_path,
     output_dir,
@@ -96,18 +109,9 @@ compile_qmd_course <- function(
   )
   
   # add compil quarto profile in each directory
-  quarto_profile_copied <- file.copy(
-    from = system.file("_quarto-compil.yml", package = "squash"),
-    to = file.path(vec_qmd_dir, "_quarto-compil.yml")
+  tmp_compil_files <- add_compil_profile_and_extension(
+    vec_qmd_path = vec_qmd_path
   )
-  # skip copy and warn if it already exists
-  if (!all(quarto_profile_copied)){
-    cli_alert_warning(
-      paste0("Existing quarto compil profile (_quarto-compil.yaml) found in:\n",
-            paste0(vec_qmd_dir[!quarto_profile_copied], collapse = "\n"),
-            "\nIt will be used for qmd rendering.")
-    )
-  }
   
   # set main folder for image
   img_root_dir <- gsub("\\.html", "_img", output_html)
@@ -125,22 +129,31 @@ compile_qmd_course <- function(
     "to modify this use {.code future::plan()}"
   ))
   
+  # setup progress report with {progressr}
+  handlers("progress")
+  p <- progressor(along = vec_qmd_path)
+  
   # render each course in parallel
   render_success <- future_map_lgl(
     .x = vec_qmd_path,
-    .f = render_single_qmd,
-    img_root_dir = img_root_dir,
+    .f = \(x){
+      p(sprintf("rendering %s", basename(x)), class = "sticky")
+      render_single_qmd(x, img_root_dir = img_root_dir)
+    },
     # make random number generation reproducible
     .options = furrr_options(seed = TRUE)
-    )
-  
+  )
+
   # exit and clean if some rendering failed
   if (!all(render_success)){
     clean_rendering_files(
       dir = vec_qmd_dir,
-      present_before = file_present_before_rendering
+      present_before = file_present_before_rendering,
+      extra_files = tmp_compil_files
     )
     return(NULL)
+  } else {
+    cli_alert_success("All qmd rendered.")
   }
 
   # read html and extract slides elements
@@ -195,7 +208,8 @@ compile_qmd_course <- function(
   
   clean_rendering_files(
     dir = vec_qmd_dir,
-    present_before = file_present_before_rendering
+    present_before = file_present_before_rendering,
+    extra_files = tmp_compil_files
   )
   
   return(